+++ updates/builder_table_update_pensoft_media_logos.php
@@ -10,10 +10,18 @@ class BuilderTableUpdatePensoftMediaLogos extends Migration
     {
         Schema::table('pensoft_media_logos', function(Blueprint $table)
         {
-            $table->integer('parent_id')->nullable();
-            $table->integer('nest_left')->nullable();
-            $table->integer('nest_right')->nullable();
-            $table->integer('nest_depth')->nullable();
+            if (!Schema::hasColumn('pensoft_media_logos', 'parent_id')) {
+                $table->integer('parent_id')->nullable();
+            }
+            if (!Schema::hasColumn('pensoft_media_logos', 'nest_left')) {
+                $table->integer('nest_left')->nullable();
+            }
+            if (!Schema::hasColumn('pensoft_media_logos', 'nest_right')) {
+                $table->integer('nest_right')->nullable();
+            }
+            if (!Schema::hasColumn('pensoft_media_logos', 'nest_depth')) {
+                $table->integer('nest_depth')->nullable();
+            }
         });
     }
 
@@ -21,10 +29,18 @@ class BuilderTableUpdatePensoftMediaLogos extends Migration
     {
         Schema::table('pensoft_media_logos', function(Blueprint $table)
         {
-            $table->dropColumn('parent_id');
-            $table->dropColumn('nest_left');
-            $table->dropColumn('nest_right');
-            $table->dropColumn('nest_depth');
+            if (Schema::hasColumn('pensoft_media_logos', 'parent_id')) {
+                $table->dropColumn('parent_id');
+            }
+            if (Schema::hasColumn('pensoft_media_logos', 'nest_left')) {
+                $table->dropColumn('nest_left');
+            }
+            if (Schema::hasColumn('pensoft_media_logos', 'nest_right')) {
+                $table->dropColumn('nest_right');
+            }
+            if (Schema::hasColumn('pensoft_media_logos', 'nest_depth')) {
+                $table->dropColumn('nest_depth');
+            }
         });
     }
 }
