+++ updates/builder_table_update_pensoft_media_videos_4.php
@@ -10,8 +10,12 @@ class BuilderTableUpdatePensoftMediaVideos4 extends Migration
     {
         Schema::table('pensoft_media_videos', function(Blueprint $table)
         {
-            $table->integer('country_id')->nullable()->default(3);
-            $table->dropColumn('country');
+            if (!Schema::hasColumn('pensoft_media_videos', 'country_id')) {
+                $table->integer('country_id')->nullable()->default(3);
+            }
+            if (Schema::hasColumn('pensoft_media_videos', 'country')) {
+                $table->dropColumn('country');
+            }
         });
     }
 
@@ -19,8 +23,12 @@ class BuilderTableUpdatePensoftMediaVideos4 extends Migration
     {
         Schema::table('pensoft_media_videos', function(Blueprint $table)
         {
-            $table->dropColumn('country_id');
-            $table->integer('country')->nullable();
+            if (Schema::hasColumn('pensoft_media_videos', 'country_id')) {
+                $table->dropColumn('country_id');
+            }
+            if (!Schema::hasColumn('pensoft_media_videos', 'country')) {
+                $table->integer('country')->nullable();
+            }
         });
     }
 }
